added a few new tests and removed a few old ones

+++ tests/tcp-client-vs-stdout.sh
@@ -12,7 +12,11 @@ SEED=$RANDOM
 
 $@ -o - --seed $SEED -n $NFILES *.l > tmp/stdout-$$
 
-$@ -o :31337 --seed $SEED -n $NFILES *.l &
+# check that we did get something
+test -s tmp/stdout-$$ || fail "didn't make anything"
+
+# this will retry connections until successful, so must be ok to start before listening
+$@ -o 127.0.0.1:31337 --seed $SEED -n $NFILES *.l &
 
 echo -n "" > tmp/tcp-$$
 
@@ -21,7 +25,7 @@ echo -n "("
 for foo in $(ol -e "(iota 0 1 $NFILES)")
 do
    echo -n "-"
-   nc localhost 31337 >> tmp/tcp-$$
+   nc -l -p 31337 >> tmp/tcp-$$
 done
 
 diff -q tmp/stdout-$$ tmp/tcp-$$ || fail "tcp server output differs from stdout output"
